fastapi-crud/requirements.txt
anyio==3.5.0
asgiref==3.5.0
black==22.1.0
click==8.0.4
fastapi==0.74.0
flake8==4.0.1
greenlet==1.1.2
h11==0.13.0
httptools==0.3.0
idna==3.3
mccabe==0.6.1
mypy-extensions==0.4.3
pathspec==0.9.0
pipdeptree==2.2.1
platformdirs==2.5.1
pycodestyle==2.8.0
pydantic==1.9.0
pyflakes==2.4.0
python-dotenv==0.19.2
PyYAML==6.0
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.31
SQLAlchemy-Utils==0.38.2
starlette==0.17.1
tomli==2.0.1
typing_extensions==4.1.1
uvicorn==0.17.5
uvloop==0.16.0
watchgod==0.7
websockets==10.1
